Adds CaseStudyMockUp module
This commit adds the `CaseStudyMockUp` module and does a little refactoring
on the `Orb` class. Three optional parameters were added to the `Orb`.
The first two are an `originXGetter` and `originYGetter` that calculate
the new origin during `Orb` init or window resize. If a value isn't given,
a default getter is used instead.

A `radiusRange` parameter was also added. This one takes two numbers, a
min and max range, and uses them to get a random radius on `Orb` init.