Add IconShowcase module
This commit adds the `IconShowcase` module, which takes a list of Icon
components. If a `typeVariant` prop is set to `true`, the module will
render a `TypeBlock`, otherwise it will render an `IconBlock`.

The `TypeBlock` converts the name of the Icon to a className, which
assigns the name as a `font-family`.

A few scrolling animations were also added using the already installed
`react-scroll-trigger`.