Adds Selected Works module
This commit adds the `SelectedWorks` module and updates the `CoreLayout`
component to render the basic front page layout. A `Header` component was
created but a navigation menu will need to be added next.

Each item in the selected works list has an `onMouseOver` event handler
that blurs other items when the item is hovered. The wrapping `<section>`
has an `onMouseOver` handler as well that runs when the user stops
hovering an item.