Add the hero module
This commit adds the hero module and stitches it up to the Contentful
data. Originally, we were going to use frontmatter to store the hero
headline, byline, etc - however with the hero being used on three pages,
I decided to create a Contentful data type for it. Especially since its
pretty configurable.

The orb container can be toggled via Hero module in Contentful. Each
Hero module slug is used to append a bespoke CSS class to the `section`
element of the component. This allows us to write CSS styles to target
specific hero modules (for instance Careers or Front Page).

The module is also configured to allow html markup in the `headline`.
Wrapping text in a `span` element will apply the knockout styles.