Add the Contact Page template
This commit adds the contact page template. In doing so we've added some
logic in the `gatsby-node` file to extract the template value from the
`layout.template` and use that to determine which Gatsby template to
render the page with.

This commit also updates the background color of the footer to match the
mockups.