Fix orb performance
This commit stops orbs from lingering in memory after the `OrbComponent`
unmounts. It does this by assigning the `orbs` array  to a ref and manually
replacing it with an empty array on unmount. It also destroys the PIXI
app afterwards, which seems to remove whatever performance impacting
elements were left over after running `stop()`.