Fixes menu animation
This commit fixes `MenuToggle`'s animations. There are two copies of
`MenuToggle`, one in `Header` and the other in `Navigation`. `Navigation`
overlays the `Header` when the mobile menu is open, and since the mobile
menu is set to `display: none` when closed, there was no initial state to
transition from when opening. This commit fixes that by handling the
menu's display using `visibility` and `opacity`.

This commit also makes slight adjustments to fix the cross's positioning
relative to the square dot menu.